Update offer button state on account change
Add updateButtonDisabledState() to onPaymentAccountSelected() in
TakeOfferViewModel since changing the account dropdown in the take
offer screen does not trigger an update of the "Next Step" button.

The default amount of an offer should not change to less than what is
actually offered, change onPaymentAccountSelected() in
TakeOfferDataModel to take into account the min. amount offered.

In TakeOfferView, the payment account defaults to the last one
selected.  Make sure TakeOfferViewModel and TakeOfferDataModel take
this into account in a couple locations.