Fix CurrencyListTest when locale is not en_US
When declaring a fiat currency and only providing the currency code,
it uses the system locale to get the currency display name.
This potentially causes different sorting results in
testUpdateWhenNotSortNumerically when locale is not en_US.

Now, it specifies the en_US locale when defining the fiat currencies
for this test.

Fixes https://github.com/bisq-network/bisq/issues/2259